updated activity feed query
